Improve doc: `to_dask_array(lengths=True)` is the key part
Clarify that it's the `lengths=True` keyword passed to `to_dask_array` that triggers computation of the chunk sizes and allows the message about 'unknown chunk size' to be addressed.

The current statement is that `to_dask_array()` doesn't work, but `to_dask_array()` can fix the problem.  That doesn't literally make sense and isn't what's meant.  I think the intent of the sentence is to introduce the `lengths=True` option and say that `to_dask_array(lengths=True)` can solve the problem, when a simple `to_dask_array()` does not.